requirements.txt
# Core
browser-use==0.9.7
playwright>=1.40.0
fake-useragent>=1.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# API Server (Phase 1)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
httpx>=0.26.0

# LLM Integration (Phase 2)
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langgraph>=0.0.20

# Vector Database (Phase 3)
chromadb>=0.4.22
qdrant-client>=1.7.0
openai>=1.10.0  # For embeddings

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
